Cherry-picks for the 4.1 branch (future 4.1.2) - 2nd batch

+++ drivers/gles3/storage/config.h
@@ -58,14 +58,14 @@ public:
 	bool use_nearest_mip_filter = false;
 	bool use_depth_prepass = true;
 
-	int max_vertex_texture_image_units = 0;
-	int max_texture_image_units = 0;
-	int max_texture_size = 0;
-	int max_viewport_size[2] = { 0, 0 };
-	int max_uniform_buffer_size = 0;
-	int max_renderable_elements = 0;
-	int max_renderable_lights = 0;
-	int max_lights_per_object = 0;
+	int64_t max_vertex_texture_image_units = 0;
+	int64_t max_texture_image_units = 0;
+	int64_t max_texture_size = 0;
+	int64_t max_viewport_size[2] = { 0, 0 };
+	int64_t max_uniform_buffer_size = 0;
+	int64_t max_renderable_elements = 0;
+	int64_t max_renderable_lights = 0;
+	int64_t max_lights_per_object = 0;
 
 	// TODO implement wireframe in OpenGL
 	// bool generate_wireframes;
